restructuring inheritance
